requirements.txt
# FoundrAI — pip requirements for Cloud Run Docker image
# Generated from environment.yml pip section.
# AutoGluon excluded (too large for Cloud Run free tier — use FLAML only in prod).
# Dev tools (black, ruff, pre-commit) excluded from production image.
#
# Install: pip install -r requirements.txt

# API framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI agents
pyautogen==0.2.28
groq==0.9.0

# RAG
haystack-ai==2.3.1
ragatouille==0.0.8.post2
sentence-transformers==2.7.0
pgvector==0.2.5

# News ingestion
newscatcherapi==0.7.0
news-please==1.5.32
apscheduler==3.10.4

# AutoML & simulation
flaml[automl]==2.1.2
numpy==1.26.4
scipy==1.13.0
pandas==2.2.2
scikit-learn==1.4.2

# Storage
supabase==2.4.3

# HTTP / async utilities
httpx==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1

# Runtime env loading
python-dotenv==1.0.1

# Testing (included for CI test stage in Docker multi-stage build)
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
faker==25.0.0
